backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
celery
redis
python-multipart
python-dotenv
pydantic[email]
clerk-backend-api
stripe
pyjwt
cryptography
